+++ AlchemistCode/Cards/Basic/Jab.cs
@@ -1,4 +1,5 @@
 using BaseLib.Utils;
+using MegaCrit.Sts2.Core.Commands;
 using MegaCrit.Sts2.Core.Entities.Cards;
 using MegaCrit.Sts2.Core.GameActions.Multiplayer;
 using MegaCrit.Sts2.Core.Models.Powers;
@@ -11,12 +12,17 @@ public class Jab : AlchemistCard
 {
     public Jab() : base(1, CardType.Attack, CardRarity.Basic, TargetType.AnyEnemy)
     {
-        WithCalculatedDamage(6, static (card, _) => Dose(card), ValueProp.Move, 3);
+        WithCalculatedDamage(5, static (card, _) => Dose(card), ValueProp.Move, 2);
+        WithVar("Poison", 2, 0);
         WithTip(typeof(PoisonPower));
     }
 
     protected override async Task OnPlay(PlayerChoiceContext choiceContext, CardPlay play)
     {
         await CommonActions.CardAttack(this, play, vfx: HitVfx("vfx/vfx_dramatic_stab")).Execute(choiceContext);
+        if (play.Target is not { IsAlive: true } target) return;
+        PoisonSplash(target);
+        await PowerCmd.Apply<PoisonPower>(choiceContext, target,
+            DynamicVars["Poison"].IntValue, Owner.Creature, this);
     }
 }